--- a/pkg/oci/remote/referrers.go
+++ b/pkg/oci/remote/referrers.go
@@ -26,5 +26,9 @@ func Referrers(d name.Digest, artifactType string, opts ...Option) (*v1.IndexMan
 	o := makeOptions(name.Repository{}, opts...)
 	rOpt := o.ROpt
 	rOpt = append(rOpt, remote.WithFilter("artifactType", artifactType))
-	return remote.Referrers(d, rOpt...)
+	idx, err := remote.Referrers(d, rOpt...)
+	if err != nil {
+		return nil, err
+	}
+	return idx.IndexManifest()
 }