#### Summary

Continuing https://github.com/sigstore/cosign/pull/2901

Previously, every call to a remote.Function was performing a Ping and Token exchange, which meant that we were doing 3 round trips instead of 1 (for most calls).

The remote.Reuse option effectively tells the remote package to call methods on a Pusher/Puller instance rather than re-doing the Ping and Token requests.

The alternative to this PR would be to plumb around a puller and pusher instance and refactor all of cosign to call methods on those, but this is much simpler.

For an empty destination repository, `cosign copy` improved from 24s to 19s. When re-running it against a repository we already pushed to, we see improvement from ~12s to ~3.5s.

| Command | Mean [s] | Min [s] | Max [s] |
|:---|---:|---:|---:|
| `cosign copy gcr.io/projectsigstore/cosign:v1.13.0 (before)` | 11.332 ± 0.212 | 11.175 | 11.876 |
| `cosign copy gcr.io/projectsigstore/cosign:v1.13.0 (after)` | 3.405 ± 0.285 | 2.819 | 3.751 |

#### Release Note

`cosign` registry operations were optimized to be much faster.

#### Documentation